Using one delete dialog for all components (#61)
