+++ client/src/components/delete-dialog.tsx
@@ -9,47 +9,26 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import { useRouter } from "next/navigation";
 
-const DeleteWorkspaceDialog = ({ workspaceId }: { workspaceId: string }) => {
-  const router = useRouter();
-  const handleSubmit = () => {
-    const jwt = localStorage.getItem("jwt");
-    const requestOptions = {
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${jwt}`,
-      },
-    };
-
-    fetch(`http://localhost:8000/api/workspaces/${workspaceId}`, requestOptions)
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error("Network response was not ok");
-        }
-
-        return response.json();
-      })
-      .then(() => {
-        router.replace("/home");
-      })
-      .catch((error) => {
-        console.error("Error deleting workspace:", error);
-      });
-  };
-
+const DeleteDialog = ({
+  componentName,
+  handleDelete,
+}: {
+  componentName: string;
+  handleDelete: () => void;
+}) => {
   return (
     <Dialog>
       <DialogTrigger>
         <div className="inline-flex items-center justify-center rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-red-700	 text-primary-foreground hover:bg-red-800 h-10 px-4 py-2">
-          Delete Workspace
+          {`Delete ${componentName}`}
         </div>
       </DialogTrigger>
       <DialogContent>
         <DialogHeader>
-          <DialogTitle>Delete Workspace</DialogTitle>
+          <DialogTitle>{`Delete ${componentName}`}</DialogTitle>
           <DialogDescription className="pt-[20px]">
-            Are you sure you want to delete this workspace?
+            {`Are you sure you want to delete this ${componentName.toLowerCase()}? This action cannot be undone.`}
           </DialogDescription>
         </DialogHeader>
         <DialogFooter className="flex justify-between w-full pt-[20px]">
@@ -63,7 +42,7 @@ const DeleteWorkspaceDialog = ({ workspaceId }: { workspaceId: string }) => {
               className="w-1/2 mr-2"
               variant="destructive"
               type="submit"
-              onClick={handleSubmit}
+              onClick={handleDelete}
             >
               Delete
             </Button>
@@ -74,4 +53,4 @@ const DeleteWorkspaceDialog = ({ workspaceId }: { workspaceId: string }) => {
   );
 };
 
-export default DeleteWorkspaceDialog;
+export default DeleteDialog;
